+++ src/client/components/GridSelection.css
@@ -16,14 +16,12 @@
 
 .selection-category {
   margin: 20px auto;
-  height: 150px;
   width: 100%;
 }
 
 .selection-category-content {
   display: flex;
-  height: 100%;
-  width: 100%;
+  height: 150px;
   box-shadow: 2px 2px 5px -5px #262626;
   background-color: whitesmoke;
   border-radius: 7px;
