+++ src/client/pages/Stories.tsx
@@ -5,7 +5,7 @@ import React from "react";
 // import React, { useEffect, useState } from "react";
 
 import Scaffold from "../components/Scaffold";
-import Selection from "../components/Selection";
+import Selection from "../components/GridSelection";
 import { View } from "../constants";
 import { service } from "../services/content";
 import { LinkPair } from "../util";
